Reorganize nodes in less files for readability
