--- src/nodes/composite.rs
@@ -1,6 +1,0 @@
-use super::drawable::Drawable;
-
-#[derive(Debug, Clone)]
-pub struct Composite {
-    pub(crate) draw_state: Drawable,
-}
