+++ src/nodes/mod.rs
@@ -1,7 +1,4 @@
-pub mod drivers;
 pub mod node;
+pub mod node_data;
 pub mod node_tree;
-
-pub mod composite;
-pub mod drawable;
-pub mod part;
+pub mod physics;
